requirements_latest.txt
PyQt6>=6.6.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
face-recognition>=1.3.0
cmake>=3.15

